Mobile support
The spacer used to keep the banner at the proper height above the
content div was edited to make it "invisible", which really means white.
Also, the ids used for the navigation bar were put in the seperators to
hopefully make the text readable when the corresponding button is
pressed.